Refactor: make backbutton folder
